requirements.txt
hikari
hikari[speedups]
hikari-crescent
prisma
rich
dyndesign
sqlitedict
fastapi
uvicorn
pydantic